Default databse with exercises, muscles and categories imported. Secondary muscles are loaded correctly. Database updated to allow template creation and added distance as a record measurement. There's a bug concerning secondary muscles in exercise creation dialog.

+++ app/src/main/java/com/rib/progressiverecords/ui/ExerciseListScreen.kt
@@ -170,22 +170,24 @@ private fun ExerciseItem(
             color = MaterialTheme.colors.onBackground
         )
 
-        Spacer(modifier = Modifier.weight(1f))
+        if (exercise.exercise.isDefault == 0) {
+            Spacer(modifier = Modifier.weight(1f))
 
-        IconButton(onClick = { onEdit(exercise) }) {
-            Icon(
-                Icons.Filled.Edit,
-                contentDescription = stringResource(R.string.edit_exercise_icon_description),
-                tint = MaterialTheme.colors.onBackground
-            )
-        }
+            IconButton(onClick = { onEdit(exercise) }) {
+                Icon(
+                    Icons.Filled.Edit,
+                    contentDescription = stringResource(R.string.edit_exercise_icon_description),
+                    tint = MaterialTheme.colors.onBackground
+                )
+            }
 
-        IconButton(onClick = { onDelete(exercise) }) {
-            Icon(
-                Icons.Filled.Delete,
-                contentDescription = stringResource(R.string.delete_exercise_icon_description),
-                tint = MaterialTheme.colors.onBackground
-            )
+            IconButton(onClick = { onDelete(exercise) }) {
+                Icon(
+                    Icons.Filled.Delete,
+                    contentDescription = stringResource(R.string.delete_exercise_icon_description),
+                    tint = MaterialTheme.colors.onBackground
+                )
+            }
         }
     }
 }
@@ -266,7 +268,7 @@ private fun deleteExercise(
 
 fun createEmptyExercise(): ExerciseWithSecMuscle {
     return ExerciseWithSecMuscle(
-        exercise = Exercise("", 0, "", "", 0),
+        exercise = Exercise("", 0, "", ""),
         muscles = emptyList()
     )
 }
